+++ CHANGELOG.md
@@ -11,6 +11,7 @@ UNRELEASED -
 ## Fixed
 - The EC2 provider will set instance tags if given in key=val format.
 - The Vultr provider doesn't wrap the tag in brackets anymore.
+- Vultr now automatically adds private IPs to the node if -p is given.
 
 ## Added
 - Extensible CLI for to create and destroy minions, and listing regions and sizes.